refactor(engine): move interfaces from api to engine
