+++ alchemist-engine/src/main/java/it/unibo/alchemist/core/BatchedScheduler.java
@@ -9,6 +9,7 @@
 
 package it.unibo.alchemist.core;
 
+import it.unibo.alchemist.core.Scheduler;
 import it.unibo.alchemist.model.Actionable;
 
 import java.util.List;
